FlashAI_v2/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
redis==4.6.0
rq==1.14.1
python-dotenv==1.0.0
pillow==10.0.0
aiofiles==23.1.0
requests==2.31.0
numpy==1.27.0
opencv-python-headless==4.8.0.76
tqdm==4.66.1
torch==2.2.0
torchvision==0.18.0
diffusers>=0.20.0
transformers==4.40.0
ftfy==6.1.1
gTTS==2.3.1
